win: do not ignore notification from .hg and .eden
Summary:
While suppressing notification should in theory be faster than getting
notification, it however causes a couple of issues due to the fact that EdenFS
is simply not creating inodes for any of the files in these directories.
Renaming a file in and out of these directories would for instance fail, due to
the inode not being present.

Since these directories are very low traffic overall and on Linux/macOS these
are not treated specifically (well, except for .eden, but that's a different
story), let's do the same on Windows.

Reviewed By: chadaustin

Differential Revision: D22250850

fbshipit-source-id: c13ed29faedc33c98b1a30227e44afc3f2c84c89