changelog: properly handle binary extras
Summary:
In configerator, extras might be non-utf8 and plain bytes. This causes 2
issues, the first one is when extracting the file list out of the change, the
other one is when dealing with the extras directly. For the first one, only
decode the file list, and for the second one, use the surrogateescape error
handling when decoding/encoding them.

Reviewed By: quark-zju

Differential Revision: D22260068

fbshipit-source-id: 19ef0248c69ea2b75bb91e59c3cdaed869110950