http_client: make client callback return a Result
Summary: Instead of returning a bool, make the client callback return a `Result`. A new error type called `Abort` has been added which the client can return to signal that the driver should abort all active transfers and return early. This results in more idiomatic code, and gives the callback the ability to specify the reason for the abort.

Reviewed By: quark-zju

Differential Revision: D22201912

fbshipit-source-id: 46d72e28f754132a4ef30defa40c4a5d09fe8e07