http_client: add HttpClient
Summary:
Add an `HttpClient` struct that is intended as the primary API for this crate. This struct is a wrapper around libcurl's "multi" interface, which allows sending several requests concurrently. Notably, if HTTP/2 is used, this allows for potentially multiplexing
many requests to the same server over the same connection.

This code is essentially a generalized version of the `multi_request` function from the EdenAPI crate.

Reviewed By: quark-zju

Differential Revision: D22157710

fbshipit-source-id: cbd7f82555444c7d1e1932944257a6949c60f34e