configs: limit how often we query the server for configs
Summary:
Since we are querying intern for remote configs, we don't want to spam
the servers with requests if they're down. Therefore let's implement some basic
rate limiting to prevent us from querying the server too often. The default
behavior is limiting it to once every 5 minutes.

We only generate new configs once every 15 minutes, so generally this rate limit
shouldn't have any effect, but if there are errors in the generation process
it's possible for generation to happen much more frequently, so this will guard
us from hitting the server too frequently.

Reviewed By: quark-zju

Differential Revision: D22243316

fbshipit-source-id: bbccaf63da95af1edc3128f4d2047a32f90e53ba