http_client: make test cli use async interface
Summary: In order to maximize the surface area of the client that gets exercised by the `http_cli` binary, make it use the new async interface (even though it is not strictly necessary here). Since the async interface is ultimately built on top of the synchronous interface, sending requests this way will still exercise parts of the synchronous interface too.

Reviewed By: quark-zju

Differential Revision: D22242944

fbshipit-source-id: 80d73cc152d0c38673436457c1e1040e4198095e