http_client: improve header parsing
Summary: Add a new `Header` type that represents a parsed header line. Notably, `libcurl` treats both the initial status line and trailing CRLF as headers; the new struct handles these cases in a strongly-typed way. This is particularly useful when working with streaming responses, as this means that the `Receiver` can tell the status code upfront rather than waiting until the end of the request, and the `Receiver` can tell once all headers (except for trailers) have been received.

Reviewed By: quark-zju

Differential Revision: D22228632

fbshipit-source-id: 06f1a21d7af25b37269bb449a1e53237ec74490a