http_client: add support for streaming CBOR deserialization
Summary: This diff adds a new `CborStream` type which wraps a `TryStream` of bytes and attempts to deserialize the data from CBOR as it is received. This allows the caller to begin processing deserialized values while the download is still going on, which is important in situations where a server might return a large streaming response composed of many small entries (as is the case with EdenAPI).

Reviewed By: quark-zju

Differential Revision: D22280745

fbshipit-source-id: 9d7fa52e4e67cf7de6bed37c28e5e7afd69449cd