win: do not ignore errors in notification callbacks
Summary:
Errors are usually a sign that we're doing something wrong. Ignoring them can
thus lead to weird behavior, hence let's not ignore them. This may have the
potential to breaking user applications if they were doing something we're not
handling properly, but it was likely that EdenFS was broken in a subtle way. We
should be able to better understand what's wrong and fix it properly then.

Reviewed By: genevievehelsel

Differential Revision: D22250852

fbshipit-source-id: 91daa60065b5621542aee1ca3207d1d7c54f8639