edenapi_types: move JSON parsing code from make_req to edenapi_types
Summary:
Move JSON request parsing code out of `make_req` into `edenapi_types` so it can be reused in the EdenAPI test CLI (added later in the stack).

Note to reviewer: This diff look bigger than it actually is; it is mostly just cut-and-paste (recorded as a copy to preserve history).

Reviewed By: quark-zju

Differential Revision: D22305693

fbshipit-source-id: 13248fb29b2fbb705203f889f3d2fb3c1c760bed