edenapi_types: add ToJson trait
Summary: Add a `ToJson` trait as a counterpart to the `FromJson` trait introduced in the last diff. The primary use case for this will be to allow recording and replaying the data fetches that occur during Mercurial operations. A JSON representation was chosen so that the format will be directly compatible with tools like `make_req` used in EdenAPI integration tests.

Reviewed By: markbt

Differential Revision: D22344599

fbshipit-source-id: 52c888bde93a8e86b6dd76cb862337f716b007eb