test-strip-cross: turn on visibility and narrow-heads
Summary:
This test works fine except for stripping rev 0 (the first commit).  Since
strip is not used in production for end-users and is likely going away, and we
are unlikely to strip the first commit of a repo on server-side, I just removed
the problematic test case to make it compatible.

Reviewed By: DurhamG

Differential Revision: D22174993

fbshipit-source-id: f362c6209852015558924151b58dc06b16c95df0