repo: narrow-heads repos are not publishing
Summary:
Usually server-side repos are not using narrow-heads. In some cases they do
(ex. push / pull without using the ssh protocol in tests). When a narrow-heads
repo receives a push without a bookmark, the repo will not update its visible
heads, causing the pushed commits to be invisible during discovery.

Change it so when narrow-heads repo receives pushes, update visible heads so
commits are visible during discovery. This is mostly for test compatibility
and does not affect production use-cases.

Reviewed By: DurhamG

Differential Revision: D22320090

fbshipit-source-id: 14625b9db0dbea6327af33bbd352a866cb10e397